Started making some good progress on the mod API
I've realized that what I need to do is create an example project for someone to use alongside the API. I think this makes the most sense. Make the API for them to depend on + interact with, then give them the example project with the default build scripts to create the .jar. After that it's on the modder to use the example project and build scripts to create a valid .jar